Update app.JS
Additional notes on "this" metod, and one interesting feature is "this" can be used as specific selection without having any unique "id" for each data, making select-edit-delete a little bit easier. The downside is, as there are no unique "id", we did not know which one of the data is being manipulated, and harder to trace back for future demands. Nevertheless, it is still easy and stands as proof of concept for this lesson.